refactor: OptionItem, OptionGroup 레포지토리 통합
- question 과 template 패키지에 각각 존재하던 것을 question 패키지로 통합
