+++ backend/src/main/java/reviewme/question/repository/OptionItemRepository.java
@@ -5,11 +5,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import reviewme.question.domain.OptionItem;
 import reviewme.question.domain.OptionType;
+import reviewme.template.domain.exception.OptionItemNotFoundException;
 
 @Repository
 public interface OptionItemRepository extends JpaRepository<OptionItem, Long> {
 
     List<OptionItem> findAllByOptionType(OptionType optionType);
 
+    List<OptionItem> findAllByOptionGroupId(long optionGroupId);
+
     boolean existsByOptionTypeAndId(OptionType optionType, long id);
+
+    default OptionItem getOptionItemById(long id) {
+        return findById(id).orElseThrow(() -> new OptionItemNotFoundException(id));
+    }
 }
